Fixed border problem
Fixed the border issue so that the computer screen size doesn't matter. Did this with Mathf.Clamp, so I got rid of some of the boundary walls. Made a trigger on the left wall so you lose if you don't keep moving foward. Can't go off the screen on the right side.